feat: reccomendations, тесты postman'а проходит
