+++ src/main/resources/application.properties
@@ -1,5 +1,3 @@
-logging.level.ru.yandex.practicum.filmorate.service.UserService=debug
-
 spring.sql.init.mode=always
 spring.datasource.url=jdbc:h2:file:./db/filmorate
 spring.datasource.driverClassName=org.h2.Driver
